referlut-api/requirements.txt
fastapi
uvicorn
nordigen
supabase
openai
httpx
beautifulsoup4
python-dotenv 